Move WebSocket package
