+++ pkg/ws/client_conn.go
@@ -1,4 +1,4 @@
-package wsconn
+package ws
 
 import (
 	"time"
