+++ pkg/ws/msg_handler.go
@@ -1,4 +1,4 @@
-package wsconn
+package ws
 
 import (
 	"fmt"
